refactor: make deviation windows cross-platform
