+++ internal/apm/deviations_stats_test.go
@@ -14,164 +14,6 @@ func completeEvidence(points int) MetricEvidence {
 	return MetricEvidence{ObservedPoints: points, ExpectedPoints: points, Coverage: 1}
 }
 
-func TestDeviationWindowResolverAvailableOnAllPlatforms(t *testing.T) {
-	now := time.Date(2026, 7, 11, 10, 7, 32, 0, time.UTC)
-	if _, err := resolveDeviationWindows(DeviationArgs{}, now, time.Minute); err != nil {
-		t.Fatal(err)
-	}
-}
-
-func TestResolveDeviationWindowsPreviousPeriod(t *testing.T) {
-	now := time.Date(2026, 7, 11, 10, 7, 32, 0, time.UTC)
-	got, err := resolveDeviationWindows(DeviationArgs{}, now, time.Minute)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got.TimeSource != "default" || got.BaselineMode != "previous_period" {
-		t.Fatalf("unexpected provenance: %+v", got)
-	}
-	if got.RequestedCurrentEnd.Sub(got.RequestedCurrentStart) != time.Hour {
-		t.Fatalf("default window duration = %s, want 1h", got.RequestedCurrentEnd.Sub(got.RequestedCurrentStart))
-	}
-	if got.EffectiveCurrentStart != time.Date(2026, 7, 11, 9, 8, 0, 0, time.UTC) ||
-		got.EffectiveCurrentEnd != time.Date(2026, 7, 11, 10, 7, 0, 0, time.UTC) {
-		t.Fatalf("unexpected effective current window: %+v", got)
-	}
-	if got.EffectiveBaselineStart != time.Date(2026, 7, 11, 8, 8, 0, 0, time.UTC) ||
-		got.EffectiveBaselineEnd != time.Date(2026, 7, 11, 9, 7, 0, 0, time.UTC) {
-		t.Fatalf("baseline offsets were not mapped: %+v", got)
-	}
-	if got.ExcludedCurrentPoints != 1 || got.ExcludedBaselinePoints != 1 {
-		t.Fatalf("excluded counts = current:%d baseline:%d", got.ExcludedCurrentPoints, got.ExcludedBaselinePoints)
-	}
-	if got.EffectiveCurrentStart.Before(got.RequestedCurrentStart) {
-		t.Fatalf("effective start moved earlier than requested: %+v", got)
-	}
-}
-
-func TestResolveDeviationWindowsRelativeLookback(t *testing.T) {
-	now := time.Date(2026, 7, 11, 10, 7, 32, 0, time.UTC)
-	got, err := resolveDeviationWindows(DeviationArgs{LookbackMinutes: 15}, now, time.Minute)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got.TimeSource != "relative_lookback" {
-		t.Fatalf("time source = %q", got.TimeSource)
-	}
-	if got.RequestedCurrentStart != now.Add(-15*time.Minute) || got.RequestedCurrentEnd != now {
-		t.Fatalf("unexpected requested current window: %+v", got)
-	}
-	if got.EffectiveCurrentStart != time.Date(2026, 7, 11, 9, 53, 0, 0, time.UTC) ||
-		got.EffectiveCurrentEnd != time.Date(2026, 7, 11, 10, 7, 0, 0, time.UTC) {
-		t.Fatalf("unexpected effective current: %+v", got)
-	}
-	if got.EffectiveBaselineStart != time.Date(2026, 7, 11, 9, 38, 0, 0, time.UTC) ||
-		got.EffectiveBaselineEnd != time.Date(2026, 7, 11, 9, 52, 0, 0, time.UTC) {
-		t.Fatalf("unexpected effective baseline: %+v", got)
-	}
-}
-
-func TestResolveDeviationWindowsExplicitCurrentRange(t *testing.T) {
-	now := time.Date(2026, 7, 11, 10, 7, 32, 0, time.UTC)
-	got, err := resolveDeviationWindows(DeviationArgs{
-		StartTimeISO: "2026-07-11T07:00:00Z",
-		EndTimeISO:   "2026-07-11T08:00:00Z",
-	}, now, time.Minute)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got.TimeSource != "explicit" || got.BaselineMode != "previous_period" {
-		t.Fatalf("unexpected provenance: %+v", got)
-	}
-	if got.EffectiveCurrentStart != time.Date(2026, 7, 11, 7, 0, 0, 0, time.UTC) ||
-		got.EffectiveCurrentEnd != time.Date(2026, 7, 11, 8, 0, 0, 0, time.UTC) {
-		t.Fatalf("historical completed window changed: %+v", got)
-	}
-}
-
-func TestResolveDeviationWindowsExplicitBaseline(t *testing.T) {
-	now := time.Date(2026, 7, 11, 10, 7, 32, 0, time.UTC)
-	got, err := resolveDeviationWindows(DeviationArgs{
-		StartTimeISO:     "2026-07-11T07:00:00Z",
-		EndTimeISO:       "2026-07-11T08:00:00Z",
-		BaselineStartISO: "2026-07-10T07:00:00Z",
-		BaselineEndISO:   "2026-07-10T08:00:00Z",
-	}, now, time.Minute)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got.BaselineMode != "explicit" {
-		t.Fatalf("baseline mode = %q", got.BaselineMode)
-	}
-	if got.EffectiveBaselineStart != time.Date(2026, 7, 10, 7, 0, 0, 0, time.UTC) ||
-		got.EffectiveBaselineEnd != time.Date(2026, 7, 10, 8, 0, 0, 0, time.UTC) {
-		t.Fatalf("unexpected effective baseline: %+v", got)
-	}
-}
-
-func TestResolveDeviationWindowsRejectsPartialPairs(t *testing.T) {
-	now := time.Date(2026, 7, 11, 10, 7, 32, 0, time.UTC)
-	for _, args := range []DeviationArgs{
-		{StartTimeISO: "2026-07-11T07:00:00Z"},
-		{EndTimeISO: "2026-07-11T08:00:00Z"},
-		{BaselineStartISO: "2026-07-10T07:00:00Z"},
-		{BaselineEndISO: "2026-07-10T08:00:00Z"},
-	} {
-		if _, err := resolveDeviationWindows(args, now, time.Minute); err == nil {
-			t.Fatalf("expected partial pair error for %+v", args)
-		}
-	}
-}
-
-func TestResolveDeviationWindowsRejectsUnequalDurations(t *testing.T) {
-	now := time.Date(2026, 7, 11, 10, 7, 32, 0, time.UTC)
-	args := DeviationArgs{
-		StartTimeISO:     "2026-07-11T07:00:00Z",
-		EndTimeISO:       "2026-07-11T08:00:00Z",
-		BaselineStartISO: "2026-07-10T07:00:00Z",
-		BaselineEndISO:   "2026-07-10T09:00:00Z",
-	}
-	if _, err := resolveDeviationWindows(args, now, time.Minute); err == nil {
-		t.Fatal("expected unequal-duration baseline error")
-	}
-}
-
-func TestResolveDeviationWindowsRejectsUnalignedHistoricalBoundaries(t *testing.T) {
-	now := time.Date(2026, 7, 11, 10, 7, 32, 0, time.UTC)
-	for _, args := range []DeviationArgs{
-		{StartTimeISO: "2026-07-11T07:00:30Z", EndTimeISO: "2026-07-11T08:00:30Z"},
-		{
-			StartTimeISO:     "2026-07-11T07:00:00Z",
-			EndTimeISO:       "2026-07-11T08:00:00Z",
-			BaselineStartISO: "2026-07-10T07:00:30Z",
-			BaselineEndISO:   "2026-07-10T08:00:30Z",
-		},
-	} {
-		if _, err := resolveDeviationWindows(args, now, time.Minute); err == nil {
-			t.Fatalf("expected unaligned historical boundary error for %+v", args)
-		}
-	}
-}
-
-func TestResolveDeviationWindowsAlignsBothWindowsToCompletedBuckets(t *testing.T) {
-	now := time.Date(2026, 7, 11, 10, 7, 32, 0, time.UTC)
-	got, err := resolveDeviationWindows(DeviationArgs{LookbackMinutes: 10}, now, time.Minute)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got.EffectiveCurrentStart != time.Date(2026, 7, 11, 9, 58, 0, 0, time.UTC) ||
-		got.EffectiveCurrentEnd != time.Date(2026, 7, 11, 10, 7, 0, 0, time.UTC) {
-		t.Fatalf("unexpected effective current window: %+v", got)
-	}
-	if got.EffectiveBaselineStart != time.Date(2026, 7, 11, 9, 48, 0, 0, time.UTC) ||
-		got.EffectiveBaselineEnd != time.Date(2026, 7, 11, 9, 57, 0, 0, time.UTC) {
-		t.Fatalf("unexpected effective baseline window: %+v", got)
-	}
-	if got.QueryStep != time.Minute {
-		t.Fatalf("query step = %s", got.QueryStep)
-	}
-}
-
 func TestSummarizeWindowUsesRatioOfTotals(t *testing.T) {
 	got := summarizeWindow([]bucket{{Requests: 1, Errors: 1}, {Requests: 999, Errors: 0}}, time.Minute, 2)
 	if got.ErrorPercentage != 0.1 {
