--- internal/apm/deviations_windows.go
@@ -1,5 +1,0 @@
-package apm
-
-// This filename is retained for task ownership history. Go treats the
-// "_windows.go" suffix as Windows-only, so cross-platform window resolution
-// lives in deviations_types.go.
