--- internal/apm/deviations_windows_test.go
@@ -1,5 +1,0 @@
-package apm
-
-// Go treats this task-mandated filename as Windows-only. Cross-platform window
-// regression tests live in deviations_stats_test.go so Linux and macOS CI run
-// the same assertions.
